renamed tasks files and folders to students to fit current project.
